Its too late, ive done too much. Split api, split login, try to purchase items
